Accelerated the khrn_hashword() function.
This function has been reimplemented in assembly, so that the data to be
processed can be preloaded ahead of when it is needed. To suit the quirks
of the BCM2708 L2 cache, each 32-byte cache line is preloaded only once,
and only at its lowest address. Speed improvements of up to 2.2x have been
measured over the un-optimised code.