Userland issue #76
Added  option to set the exposure/shutter time.

Needs a change on GPU binary, Brcm ref: 425060 and
a consequent change to
userland/interface/mmal/mmal_parameters_camera.h

There is currently a limitation of about 350ms beyond which
the camera will lock (depending on other settings). This
is being investigated, but is unrelated to this change
i.e. You can get the fault in other ways, not just by setting
the shutter speed.