Add circular buffer mode.
This commit adds the functionality to set up a circular buffer in memory
which stores the last N seconds of video.  When the application is send a
signal or a key is pressed the last N seconds are saved.

Currently it seems to work well enough although there seem to be problems
with playing back some video using omxplayer when the GOP size is changed.