Add MMAL callbacks for various camera parameters
Add MMAL support for OMX_IndexConfigCameraSettings which
allows reporting of parameters such as exposure, analog gain,
digital gain, AWB values, and focus position back to the app
as they change.

Not quite complete as this requires the port to be the control
port (OMX_ALL), and that uses the video Camplus stats. You can
manually request the values from the stills port, if desired.

NB Updates both MMAL and IL headers