vcos: Fixed include for "HZ"
There is no guarantee that <sys/param.h> defines HZ. And in musl libc,
it doesn't. Since this is a Linux-specific constant, include the
Linux-specific header for it.

Note that HZ as defined by the system headers is the default value of
HZ (100), which might differ from the actual value of HZ in the kernel
config. Ideally userland would not use HZ at all and do all timing in
milliseconds instead.

Signed-off-by: Maarten ter Huurne <maarten@treewalker.org>