mmal: allow non-tunnelled clock port connections
Clock connections now behave in a similar way to tunnelled clock
ports in that buffers received from one clock port are sent directly
to the other clock port without going through the connection queue.

This does mean, however, that a client won't be able to intercept
clock buffers and that a component shouldn't perform significant
processing in the clock port callback.

mmal-graph: add support for clock ports

Allow clock ports to be added to a graph component.

mmal: only send buffers to enabled clock ports

When using a non-tunnelled connection between clock ports, the buffer
header callback should only attempt to forward a buffer if the other
clock port in the connection is enabled.  This avoids errors being
logged when disabling a clock connection.