mmal: don't use bulk transfers for clock buffers
Clock buffers only contain a small amount of data which can fit into the
vchiq control message, so use that instead of doing a bulk transfer.

mmal/vc: fix mem leak when using opaque buffers

When in opaque mode, we would leak the payloads if
the zero copy mode was not enabled.
Also fix an issue where if somehow the format of the port has
changed by the time buffers are being freed, we end up not
releasing them.