Example code for storing a FBO in VC shared memory
Add a new RaspiStill example scene (vcsm_square) that shows how to
render a camera buffer using a GLES2 shader to an frame-buffer texture
in shared memory. The texture is then modified by the CPU and drawn to
the window surface.

The example just draws an animated rectangle to demonstrate that the
pipeline is working. A more interesting example might use a GL
shader to do edge detect, pass the output to OpenVC and then render
or composite the result.

To run 'raspistill --gl --glscene vcsm_square'