Add EGL client support for EGL_IMAGE_BRCM_VCSM
Allocate a VCSM buffer using the dimensions supplied by the caller and
pass the vc mem-handle to 3D driver.

N.B. This API requires an update to the VideoCore firmware. If the update
is not present then eglCreateImageKHR will return an error if VCSM EGL
images are requested.

If HAVE_VCSM is defined then EGL is statically linked with libvcsm,
otherwise, libvcsm is loaded lazily as a plug-in when the first VCSM
EGL vcsm image target is requested.