vmcs_host: Poll for multiple dispmanx messages
When VCHI delivers a callback indicating that a message is available,
there could be more than one waiting. Therefore it is necessary to
poll until vchi_msg_dequeue fails.

Without this patch, failure to keep up with vsync messages will cause
a gradually increasing tailback, eventually leading to a stall because
the slots aren't being recycled.