Deprecate pthreads vcos_get_free_mem()
The pthreads version of vcos_get_free_mem() isn't used anywhere
and causes compiler warnings due to its use of sbrk() without
the relevant #defines.

Ideally it should look at something like /proc/vmstat, but
that's not what it does.

So turn it into a stub function and mark it deprecated. The
VMCS version will carry on working as before.

See issue #316.