vcos_verify() warnings - discard result from vcos_verify()
It looks like vcos_verify() is being used here to enable
better debugging (hit a breakpoint if we ever take this code
path and drop into a debugger, but in normal use do nothing).

The compiler now warns about the ignored vcos_verify() rvalue,
so just cast the result to (void).