Add ARM64 CMake variable for 64-bit targets
At the moment, much of the codebase compiles on 64-bit targets, but not
all (mostly 32-bit assembly).  -DARM64=ON allows for skipping
directories which currently do not compile correctly.

This is backwards compatible with existing behavior (i.e. no -DARM on a
32-bit target compiles everything as before), and hopefully will
ultimately become unnecessary as remaining components are fixed for
64-bit use.