vchiq_test: Fix race condition in clnt_callback
In clnt_callback in the message available path, the
message may be released before it is inspected.  This can
cause the message to not be recognized and the necessary event
signaled causing the test to hang.

The solution is to store the header size in a variable before
it is released.