dtoverlay: Update fixups when a node is renamed
When the firmware applies an overlay it processes the fixups before
applying any parameters. Under Linux, parameters are handled by the
dtoverlay/dtparam application before submission to the kernel, which
is where the fixups are applied. As a consequence, fixups must be
modified if they refer to nodes which are renamed as the result of
using a parameter.

The process of updating the fixups (and symbols, in case the kernel
cares in the future) is too heavyweight to want the firmware to
do it unnecessarily, so internally use the fact that the fixups
have been applied as an indication that they no longer need to be
patched.

See: https://github.com/raspberrypi/linux/issues/2002