dtoverlay app: Add the -D (dry-run) option
Dry-run mode does the same processing as adding an overlay, but instead
of submitting the overlay to the kernel it saves the result in
"dry_run.dtbo" in the current directory. This feature is provided as a
diagnostic aid - when combined with the dtmerge command it allows one
to simulate the application of an overlay in the same manner as the
dynamic overlays are applied.