Camera/tuner: Allow client to override lens shading table.
Provide a mechanism to push an amended lens shading table
for when the user has changed the optics significantly.
It disables all ALCS and other manipulation of the table.

Provided table copies part of the OMAP3 style, taking a 64x64
grid of values. Each value is 3p5 fixed point representation,
same OMAP mode 6
0x6: Coded as 3-bit integer, 5-bit fraction Range from 0
to 7+31/32