MMAL: Fix example_basic_2 to actually play video
example_basic_2 seemed to be broken. It tried to use
ZERO_COPY but in an invalid way. It also never looked
for format changed events off the decoder, so would
just stop once the decoder issued those.

It now correctly allocates pools so that it can use zero
copy, and also acts on MMAL_EVENT_FORMAT_CHANGED events
to reallocate the pools and hence proceed.

Fixes #285