dtoverlay app: Keep overlay symbols private
As of 4.14, the kernel overlay merging includes any symbols defined
by an overlay, which can cause clashes between overlays (and even
within them in the case that an overlay deliberately adds a global
symbol as a way of exporting it to other overlays).

In future we may choose to make use of this mechanism as a nicer
way of exporting symbols, but for now just delete all symbols before
presenting the overlay to the kernel.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>