dtoverlay: Use SUDO_USER in -pre and -post scripts
This patch replaces the hardcoded "pi" user in the dtoverlay-pre and
dtoverlay-post scripts with $SUDO_USER, which is set by the sudo command
to the ID of the user calling it. In the event that SUDO_USER isn't set,
$LOGNAME is used (set by login shells), with an ultime fallback to "pi".
This change is necessary for users who have changed the login ID, perhaps
as a security measure.