dtoverlay: Also allow "fragment-<n>" in overlays
The old "fragment@<n>" syntax causes newer versions of dtc to complain
about missing "reg" parameters, so there is a move to replace the '@'s
with '-'s. All that really matters is that the node names are distinct.
Modify the dtoverlay library to cope with either form. I don't think
it is necessary to cope with arbitrary fragment names, and keeping
this restriction makes handling disabled ("__dormant__") fragments
easier.