MMAL: Populate buffer header TYPE_SPECIFIC fields.
MMAL_BUFFER_HEADER_T has a TYPE_SPECIFIC fields, which for
video includes the number of planes, pitches and offsets.
This hasn't been populated previously, but it can be helpful
for interfacing to the likes of DRM and FFMPEG to be given
these values.

RIL doesn't support these values, so precompute them for
output ports based on the port format.