mmal: Relax requirement on a buffer in mmal_port_send_buffer
If the alloc_size of a buffer is flagged as zero, then remove
the requirement for buffer->data to be non-NULL in
mmal_port_send_buffer.

This is initially an optimisation for the V4L2 codec support
where there is a need to send an EOS flag on an empty buffer,
so saves allocating the buffer which will never be used.