MMAL/IL: Add support for components to report alignment requirements.
Add parameter to allow RIL components to report the buffer alignment
requirements, and the core will the do those alignments.

It would make more sense for the mmal_port_format_commit to fail
should the width/height not be suitably aligned, rather than just
amending the format and not telling userland.