Fixed up initial timelapse delay
When in timelapse mode, the delay before the first shot
was the timelapse value. However, if a low value is set
this does not give enough time for the AE etc to settle.

In addition, the first image was only taken after the
first delay, rather than immediately which is what would
be expected. So a 5minute timelapse delay meant the 1st
image was only taken after 5 minutes.

The patch sets the delay for the first image capture to
be a constant value, 1000ms. So always time for the AE etc
to settle.

Fixes: #429, #473