Option to specify timestamp in the segment filename
In segmented mode, this allows you to specify that the file
name for each segment uses a time based filename, not a segment
number based one.
If the filename contains %d or %u a segment number is used.
Anything else, and the filename string is used as a formatting
string for the strftime function.