hello_pi: encode: Process the buffer before returning for filling.
The code was calling ilclient_get_output_buffer to retrieve a
filled buffer, and then immediately returning it via OMX_FillThisBuffer
before looking at the contents.
OMX_FillThisBuffer does not block, but relies on waiting for the
FillBufferDone callback instead.

The correct order of doing things is to call ilclient_get_output_buffer
(which waits on FillBufferDone if necessary), process the buffer,
and then call OMX_FillThisBuffer to pass it back to OMX for filling
again.

The recent improvements to video_encode mean that it is faster at
processing buffers and would often have reset the buffer header before
hello_encode had a chance to save the data.