mmal: Add KEEP_PORT_FORMATS flag to mmal connection
There's a slightly quirky use case with deinterlacing.
Interlaced YUV420 is line interleaved, so chroma line 1
is that for luma lines 1&3, and chroma line 2 is for luma lines 2&4.
If you pass such a frame into a standard component (eg the ISP), it
messes up the chroma by assuming chroma line 1 is for luma lines 1&2.

If you pass it into the ISP as double width by half height then the
chroma subsampling behaves correctly. Normally that precludes using
a mmal_connection as that will copy the port format from output
to input. Setting this new flag skips that stage and makes it the
client's responsibilty to set both port formats appropriately.