arm_dt: Work around an absent cache-line-size
The upstream VCHIQ has been denied its cache-line-size property and
forces a cache line size of 32. This doesn't work on V7/8 cpus with
64-bit cache lines.

If an updated DT is found (based on the size encoded in the reg
property, which should be 0x3c not 0xf, and the vchiq node name
"mailbox@7e00b840"), assume the kernel is using the correct cache
line size. A corresponding kernel patch derives the correct value,
and updates the size as indicated.