Raspivid: Fix for bug where %04d no longer works.
When inserting the time/date stamp into the filename was added
certain combinations of output filename no longer work. As it
was only searching for "%u" or "%d". Specifying %04d would take
the code down the timestamp route but with invalid parameters.

Make the search slightly more intelligent to allow specifying
the number of digits to insert in the filename.