dispmanx: Add option to ignore all layers lower than the current layer
Adds option DISPMANX_FLAGS_ALPHA_DISCARD_LOWER_LAYERS to DispmanX
alpha flags that sets the current layer as effectively totally opaque
and full screen, therefore obscuring anything below and avoiding adding
it to the HVS display list.

Plumb it through IL/MMAL as well so that it can be used as a simple way
to hide the frame buffer from the display.