Add media flushing call on when writing bitstream
Under certain circumstances, the Linux file system write behind caching
get so much data to write in one go that is causes frames to be dropped.
The current fflush is not sufficient as that only flushes the C library
caches, not the Linux filesystem caches.

This fix forces data to be written to media as soon as it arrives, using
fdatasync().

Note, currently it's conditional on the --flush  command line parameter.