dtoverlay: Plug small memory leak
The dtoverlay library will leak an allocation of 4 bytes for every
application of an overlay that attempt to set a "reg" property that
doesn't already exist, e.g. the old i2c-gpio overlay. This is an
error case since it won't do what the user expects, but that's no
excuse to leak.