isp: allow more formats of lens shading tables, including corner-sampled and 16-bit
The old lens_shading parameter structures exist for backwards
compatability, but there's a new v2 version which includes a
corner_sampled flag (when the grid is sampled on cell corners rather
than centrally) and a gain_format, which allows different fractional
precision in the uint8_t values, and even a u4.10 16-bit format.

Calls through the V4L2 interface to the ISP should use the v2
style of structure.