libdtovl: Allow labels to be exported
Give overlay writers the ability to export symbols so that they become
visible to other overlays once the overlay is applied. Symbols (labels)
for exporting should be listed as empty properties in a first-level
node called "__exports__".

The new API function dtoverlay_filter_symbols deletes non-exported
symbols, and dtoverlay_merge_overlay rewrites any exported symbols
so as to be relative to the base tree.