dtoverlay: Preserve exported labels
Up to now, all labels have been local to overlays. It has been possible
to explicitly add labels to the base DTB, but only in a way that is
ugly and fragile - it requires knowledge of the absolute path to where
the overlay will be applied (which could vary between DTBs), and would
break if the target of a fragment was changed.

Overlay writers can now choose to explicitly export some labels by
listing them as empty properties in a first-level node called
"__exports__". For example, to export a label called "newnode", add
the following node to the overlay:

    __exports__ {
        newnode;
    };

Make use of the new ability by replacing the deletion of the __symbols__
node with a call to dtoverlay_filter_symbols.