MMAL/IL/ISP component: Add support for mono MIPI formats, and Bayer raw 14
MIPI raw 14 wasn't previously defined, so add it for completeness

To support monochrome image sensors, add formats to handle the relevant
packed data formats into the ISP, and disable demosaic.
The Bayer pipeline is still active in Bayer mode, so all 4 channels need
to be programmed with the same data for blocks such as black level and
lens shading.