Remove gps.h and use distro's version
This patch removes the copy of gps.h from the code, relying instead on
the gps.h shipped with the distro (by libgps-dev).

That much is simple enough, but the current version of libgps has moved
from using doubles (as timestamp_t) to timespec_t structures for greater
accuracy. This patch also adds some trivial fixes for this change.