CMakeLists: Introduce AMD64 (to fix x86_64/--native builds)
CMakeLists.txt currently assumes everything not ARM64 to be ARM32. When
compiling on x86_64-linux, this tries to build some binaries that are
arm(32) only.

Some tools make some sense to use even on non-ARM workstations. For
example, `dtmerge` can be used to merge .dtb files with overlays on a
x86_64 machine (while mounting an sdcard for example).

This introduces a "AMD64" cmake flag, that's set in the `buildme` script
only when compiling for x86_64.

Signed-off-by: Florian Klink <flokli@flokli.de>