mmal: Hack to make mmal core register VideoCore components
The default linker option with Bullseye appears to now set
as-needed, so as the core doesn't call into mmal_vc_client
it is viewed as unnecesary, missing that mmal_vc_client has
a constructor that registers functions with the core.

Move the registration of the component supplier to the core_init.
The linker is therefore satisfied that mmal_vc_client is used,
and actually links to it.

This allows Picamera to work again.

https://github.com/waveform80/picamera/issues/697

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>